Patch 0.1.2: Moved theme file into common directory
